exercicios do dia 23

+++ app/src/main/java/br/com/caelum/cadastro/dao/AlunoDao.java
@@ -16,7 +16,7 @@ import br.com.caelum.cadastro.modelo.Aluno;
  */
 public class AlunoDao extends SQLiteOpenHelper {
 
-    private static final int VERSAO = 1;
+    private static final int VERSAO = 2;
     private static final String TABELA = "Alunos";
     private static final String DATABASE = "CadastroCaelum";
 
@@ -38,9 +38,14 @@ public class AlunoDao extends SQLiteOpenHelper {
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        String sql = "DROP TABLE IF EXISTS " + TABELA;
-        db.execSQL(sql);
-        onCreate(db);
+        if(newVersion == 1) {
+            String sql = "DROP TABLE IF EXISTS " + TABELA;
+            db.execSQL(sql);
+            onCreate(db);
+        } else if(newVersion == 2) {
+            String sql = "ALTER TABLE " + TABELA + " ADD COLUMN caminhoFoto TEXT;";
+            db.execSQL(sql);
+        }
     }
 
     public void inserir(Aluno aluno) {
@@ -50,13 +55,32 @@ public class AlunoDao extends SQLiteOpenHelper {
         values.put("endereco", aluno.getEndereco());
         values.put("site", aluno.getSite());
         values.put("nota", aluno.getNome());
+        values.put("caminhoFoto", aluno.getCaminhoFoto());
 
         getWritableDatabase().insert(TABELA, null, values);
     }
 
+    public void alterar(Aluno aluno) {
+        ContentValues values=new ContentValues();
+        values.put("nome", aluno.getNome());
+        values.put("telefone", aluno.getTelefone());
+        values.put("endereco", aluno.getEndereco());
+        values.put("site", aluno.getSite());
+        values.put("nota", aluno.getNome());
+        values.put("caminhoFoto", aluno.getCaminhoFoto());
+
+        getWritableDatabase().update(TABELA, values, "id=?", new String[] {aluno.getId().toString()});
+    }
+
+    public void apagar(Aluno aluno) {
+        SQLiteDatabase db=getWritableDatabase();
+        String[] args={aluno.getId().toString()};
+        db.delete(TABELA, "id=?", args);
+    }
+
     public List<Aluno> getLista() {
         List<Aluno> alunos=new ArrayList<>();
-        Cursor c = getReadableDatabase().rawQuery("SELECT * FROM " + TABELA + ";", null);
+        Cursor c = getReadableDatabase().rawQuery("SELECT * FROM " + TABELA + " ORDER BY NOME;", null);
 
         while(c.moveToNext()) {
             Aluno aluno=new Aluno();
@@ -67,10 +91,10 @@ public class AlunoDao extends SQLiteOpenHelper {
             aluno.setEndereco(c.getString(c.getColumnIndex("endereco")));
             aluno.setSite(c.getString(c.getColumnIndex("site")));
             aluno.setNota(c.getDouble(c.getColumnIndex("nota")));
+            aluno.setCaminhoFoto(c.getString(c.getColumnIndex("caminhoFoto")));
             alunos.add(aluno);
         }
         c.close();
         return alunos;
-
     }
 }
